Move UI components from JusPrinView3D to JusPrinPlaterOverlay
- Moved ChatActivationButton and ActivationButtonNotificationBadge classes from JusPrinView3D to JusPrinPlaterOverlay
- Updated includes to avoid circular dependencies
- Improved code organization by placing UI components with their managing class
- Simplified JusPrinView3D to focus only on 3D view functionality
