+++ src/slic3r/GUI/JusPrin/JusPrinPlaterOverlay.hpp
@@ -3,17 +3,56 @@
 
 #include <wx/panel.h>
 #include <string>
+#include <functional>
 
 // Forward declarations
 class wxWindow;
+class wxPaintEvent;
+class wxMouseEvent;
+class wxAnimationCtrlBase;
 
 namespace Slic3r {
 namespace GUI {
 
 // Forward declarations
 class JusPrinChatPanel;
-class ChatActivationButton;
-class ActivationButtonNotificationBadge;
+
+// Notification badge for the chat activation button
+class ActivationButtonNotificationBadge : public wxPanel {
+public:
+    ActivationButtonNotificationBadge(wxWindow* parent, const wxString& text, const wxColour& bgColor);
+
+    // Method to update text
+    void SetText(const wxString& text) {
+        m_text = text;
+        Refresh();
+    }
+
+private:
+    void OnPaint(wxPaintEvent&);
+    wxString m_text;
+    wxColour m_bgColor;
+};
+
+// Button to activate the chat panel
+class ChatActivationButton : public wxPanel
+{
+public:
+    ChatActivationButton(wxWindow* parent, wxWindowID id = wxID_ANY, const wxPoint& pos = wxDefaultPosition, const wxSize& size = wxDefaultSize);
+    void DoSetSize(int x, int y, int width, int height, int sizeFlags = wxSIZE_AUTO) override;
+    void AddJoin(std::function<void(wxMouseEvent&)> do_some) { m_do = do_some; }
+
+private:
+    void OnPaint(wxPaintEvent& event);
+    void OnMouseEnter(wxMouseEvent& event);
+    void OnMouseLeave(wxMouseEvent& event);
+    void OnMouseMove(wxMouseEvent& event);
+
+private:
+    bool                 m_isHovered{false};
+    wxAnimationCtrlBase* m_animationCtrl{nullptr};
+    std::function<void(wxMouseEvent&)> m_do{nullptr};
+};
 
 // Class to manage the chat panel and related UI components that display 
 // on top of the Plater in both Prepare and Preview tabs
